requirements.txt
blessings==1.7
Cerberus==1.3.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
PyYAML==5.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
